re-organize tooling to be used by explorers/mdims
